Update where things live. Update docstrings. Update tests. Remove unused repeated code.
